+ save listview state, refactorings and etc.

+++ WinServicesMgr/MainForm.cs
@@ -1,12 +1,17 @@
 ﻿using System;
+using System.IO;
 using System.Data;
 using System.Linq;
 using System.Drawing;
 using Microsoft.Win32;
 using System.Management;
+using System.Reflection;
 using System.Diagnostics;
 using System.Windows.Forms;
 using System.ServiceProcess;
+using WinServicesMgr.Helpers;
+using WinServicesMgr.Entities;
+using System.Collections.Generic;
 
 namespace WinServicesMgr
 {
@@ -17,44 +22,68 @@ namespace WinServicesMgr
         /// Servisleri ('DisplayName'-i A-dan Z-ye dogru siralanmiw bir wekilde) ozunde saxlayir.
         /// </summary>
         private static readonly ServiceController[] services = ServiceController.GetServices().OrderBy(x => x.DisplayName).ToArray<ServiceController>();
+
+        /// <summary>
+        /// ttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttt
+        /// </summary>
+        private static string DestinationFilePath = Assembly.GetExecutingAssembly().Location.Substring(0, Assembly.GetExecutingAssembly().Location.LastIndexOf(@"\")) + @"\ServicesList.json";
         #endregion Vars
 
         public MainForm()
         {
             InitializeComponent();
 
+            #region Form settings
             this.Text = "WinServicesMgr";
+            this.Icon = WinServicesMgr.Properties.Resources.app;
             this.MaximizeBox = false;
             this.FormBorderStyle = FormBorderStyle.FixedSingle;
             this.Font = new Font("Segoe UI", 12);
+            this.StartPosition = FormStartPosition.CenterScreen;
+            #endregion Form settings
 
+            #region ListView settings
             lvServices.View = View.Details;
             lvServices.GridLines = true;
             lvServices.FullRowSelect = true;
             lvServices.MultiSelect = false;
             lvServices.HeaderStyle = ColumnHeaderStyle.Nonclickable;
+            #endregion ListView settings
         }
 
         private void WinServicesMgr_Load(object sender, EventArgs e)
         {
-            foreach (ServiceController service in services)
+            if (File.Exists(DestinationFilePath))
             {
-                ListViewItem listViewItem = new ListViewItem(new string[] { service.ServiceName, service.Status.ToString(), service.DisplayName });
-                switch (service.Status)
+                if (new FileInfo(DestinationFilePath).Length > 0)
                 {
-                    case ServiceControllerStatus.Stopped:
-                    case ServiceControllerStatus.StopPending: listViewItem.BackColor = Color.OrangeRed; break;
+                    try
+                    {
+                        List<ServiceEntity> resultEntity = JsonHelper<ServiceEntity>.Deserialize(DestinationFilePath);
+                        foreach (var entity in resultEntity)
+                        {
+                            ControlHelper.AddToListViewAndBeautify(lvServices, entity.ServiceName, entity.ServiceStatus, entity.DisplayName);
+                        }
+                    }
+                    catch { lvServices.Items.Clear(); }
+                }
+            }
+            else
+            {
+                List<ServiceEntity> listOfServiceEntity = new List<ServiceEntity>();
 
-                    case ServiceControllerStatus.StartPending:
-                    case ServiceControllerStatus.Running: listViewItem.BackColor = Color.LawnGreen; break;
+                foreach (ServiceController service in services)
+                {
+                    listOfServiceEntity.Add(new ServiceEntity() { DisplayName = service.DisplayName, ServiceName = service.ServiceName, ServiceStatus = service.Status });
 
-                    case ServiceControllerStatus.PausePending:
-                    case ServiceControllerStatus.Paused: listViewItem.BackColor = Color.Yellow; break;
-
-                    default: listViewItem.BackColor = Color.White; break;
+                    ControlHelper.AddToListViewAndBeautify(lvServices, service.ServiceName, service.Status, service.DisplayName);
                 }
 
-                lvServices.Items.Add(listViewItem);
+                JsonHelper<ServiceEntity>.Serialize
+                (
+                    Entity: listOfServiceEntity,
+                    FilePath: DestinationFilePath
+                );
             }
 
             lvServices.AutoResizeColumns(ColumnHeaderAutoResizeStyle.ColumnContent);
@@ -73,8 +102,9 @@ namespace WinServicesMgr
                         if (service.ServiceName == selectedServiceName)
                         {
                             var servis = new ManagementObject(new ManagementPath(string.Format($"Win32_Service.Name='{service.ServiceName}'")));
-                            if (servis["Description"] != null) { rtbDescription.Text = servis["Description"].ToString(); }
-                            else { rtbDescription.Text = "No Description Found"; }
+                            rtbDescription.Text = servis["Description"] != null ? servis["Description"].ToString() : "No Description Found";
+
+                            rtbDescription.Text = rtbDescription.Text.TrimEnd().EndsWith(".") ? rtbDescription.Text : rtbDescription.Text.Insert(rtbDescription.Text.Length, ".");
                         }
                     });
                 }
@@ -84,19 +114,43 @@ namespace WinServicesMgr
 
         private void lvServices_MouseDoubleClick(object sender, MouseEventArgs e)
         {
-            string selectedServiceName = lvServices.SelectedItems[0].SubItems[0].Text;
+            if(e.Button == MouseButtons.Left)
+            {
+                string selectedServiceName = lvServices.SelectedItems[0].SubItems[0].Text;
 
-            /* Regedit aciq idise, birbawa acilmasini istediyimiz hedef Keye yonlendirib gostere bilmerik, cunki artiq Regedit aciqdir. Yonlendirmemiwden qabaq regediti baglamaliyiq, baglamasaq, aciq qalsa regedit, yeni acmaga caliwdigimiz regedit acilmayaq ve evvelceden aciq olan regedit penceresine fokuslanacayiq. Ona gorede regediti baglayiriq Keyi acmamiwdan qabaq: */
-            foreach (Process proc in Process.GetProcessesByName("regedit")) proc.Kill();
+                /* Regedit aciq idise, birbawa acilmasini istediyimiz hedef Keye yonlendirib gostere bilmerik, cunki artiq Regedit aciqdir. Yonlendirmemiwden qabaq regediti baglamaliyiq, baglamasaq, aciq qalsa regedit, yeni acmaga caliwdigimiz regedit acilmayaq ve evvelceden aciq olan regedit penceresine fokuslanacayiq. Ona gorede regediti baglayiriq Keyi acmamiwdan qabaq: */
+                foreach (Process proc in Process.GetProcessesByName("regedit")) proc.Kill();
 
-            /* Regedit acilanda fokuslanmagini istediyim Key: */
-            var Path = Registry.LocalMachine.OpenSubKey($"SYSTEM\\CurrentControlSet\\Services\\{selectedServiceName}");
+                /* Regedit acilanda fokuslanmagini istediyim Key: */
+                var Path = Registry.LocalMachine.OpenSubKey($"SYSTEM\\CurrentControlSet\\Services\\{selectedServiceName}");
 
-            /* Reyestrda son aciq qalan yolu saxlayan Keydir awagidaki, bu Keyde olan yol acilir Regedit proqrami acilanda. Eger bu keyin deyerini acilmasini istediyimiz Key ile deyiwsek, demeli regedit acilanda son aciq qalan Key olaraq bizim verdiyimiz Keyi bilecek: */
-            Registry.CurrentUser.CreateSubKey("Software\\Microsoft\\Windows\\CurrentVersion\\Applets\\Regedit").SetValue("LastKey", Path);
+                /* Reyestrda son aciq qalan yolu saxlayan Keydir awagidaki, bu Keyde olan yol acilir Regedit proqrami acilanda. Eger bu keyin deyerini acilmasini istediyimiz Key ile deyiwsek, demeli regedit acilanda son aciq qalan Key olaraq bizim verdiyimiz Keyi bilecek: */
+                Registry.CurrentUser.CreateSubKey("Software\\Microsoft\\Windows\\CurrentVersion\\Applets\\Regedit").SetValue("LastKey", Path);
 
-            /* Regedit-i bawlat: */
-            Process.Start("regedit");
+                /* Regedit-i bawlat: */
+                Process.Start("regedit");
+            }
+        }
+
+        private void exitToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            Application.Exit();
+        }
+
+        private void exportToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            using (SaveFileDialog SFD = new SaveFileDialog())
+            {
+
+            }
+        }
+
+        private void importToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            using (OpenFileDialog OFD = new OpenFileDialog())
+            {
+
+            }
         }
     }
 }
